Organise parsers into subpackages
