+++ src/main/java/seedu/address/logic/parser/cookbook/CookbookCommandParser.java
@@ -1,0 +1,20 @@
+package seedu.address.logic.parser.cookbook;
+
+import seedu.address.logic.commands.cookbook.CookbookCommand;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.exceptions.ParseException;
+
+/**
+* Parses input arguments and into a {@code CookbookCommand}.
+ */
+public class CookbookCommandParser implements Parser<CookbookCommand> {
+
+    /**
+     * Parses the given {@code String} of arguments in the context of a CookbookCommand
+     * and returns a CookbookCommand object for execution.
+     * @throws ParseException if the user input does not conform the expected format
+     */
+    public CookbookCommand parse(String args) throws ParseException {
+        throw new ParseException("Not implemented yet");
+    }
+}
