+++ src/main/java/seedu/address/logic/parser/inventory/InventoryCommandParser.java
@@ -1,0 +1,20 @@
+package seedu.address.logic.parser.inventory;
+
+import seedu.address.logic.commands.inventory.InventoryCommand;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.exceptions.ParseException;
+
+/**
+* Parses input arguments and into a {@code InventoryCommand}.
+ */
+public class InventoryCommandParser implements Parser<InventoryCommand> {
+
+    /**
+     * Parses the given {@code String} of arguments in the context of a InventoryCommand
+     * and returns a InventoryCommand object for execution.
+     * @throws ParseException if the user input does not conform the expected format
+     */
+    public InventoryCommand parse(String args) throws ParseException {
+        throw new ParseException("Not implemented yet");
+    }
+}
